requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Login>=0.6.2
Flask-WTF>=2.0.0
Werkzeug>=2.0.0
python-dotenv>=0.19.0
WTForms>=3.0.0
email-validator>=2.0.0
